Refaktor struktur folder
