+++ src/presidential-candidate/campaign-promises/CampaignPromisesIntro.tsx
