+++ src/presidential-candidate/campaign-promises/CandidateCampaignCard.tsx
